[Codegen] NFC: Lift InnerTiledOp unroll pattern to Codegen.
Follow-up to the previous NFC, moving the third (and last)
`iree_codegen.inner_tiled` rewrite pattern out of
`Codegen/Dialect/GPU/Transforms/`:

  - `UnrollInnerTiledPattern` (unrolls along iter dimensions until each
    intrinsic invocation has a unit iter shape)

The pattern itself is dialect-neutral; only the convenience
no-args populate function applied a default config (matmul-like
LHS-reuse traversal order, unit native shape) that happens to be a
good fit for any architecture, including CPU.

Renames the populate functions:

  - `IREE::GPU::populateIREEGPUVectorUnrollPatterns` (with-options form)
        -> `IREE::Codegen::populateUnrollInnerTiledPatterns`
  - `IREE::GPU::populateIREEGPUVectorUnrollPatterns` (no-args form)
        -> `IREE::Codegen::populateUnrollInnerTiledPatterns`

Renames the GPU-specific helper `gpuMatmulLikeUnrollOrder` to
`matmulLikeUnrollOrder` (its content is dialect-neutral). The
`kUnrollAccDistribute` / `kUnrollAccReassemble` HoistableConversionOp
tag constants move with the pattern.

Updates the in-tree callers in `IREEGPUExtensions.cpp` and
`UnrollToIntrinsics.cpp`, and adjusts BUILD.bazel/CMakeLists.txt deps.

Strict NFC; no functional or behavioral change.


Progress towards #24323
