+++ compiler/src/iree/compiler/Codegen/Dialect/GPU/Transforms/Transforms.h
@@ -187,10 +187,6 @@ void mapLaneForalls(RewriterBase &rewriter, Operation *funcOp,
 // Various populate pattern methods.
 void populateIREEGPULowerBarrierRegionPatterns(RewritePatternSet &patterns);
 void populateIREEGPULowerValueBarrierPatterns(RewritePatternSet &patterns);
-void populateIREEGPUVectorUnrollPatterns(
-    RewritePatternSet &patterns, const vector::UnrollVectorOptions &options);
-// Version of unrolling with a preset configuration.
-void populateIREEGPUVectorUnrollPatterns(RewritePatternSet &patterns);
 // Populate patterns to fold tensor.empty ops through swizzle hint ops.
 void populateFoldSwizzleHintOpPatterns(RewritePatternSet &patterns);
